Тип int в duration

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -14,5 +14,5 @@ public class Film {
     String name;
     String description;
     LocalDate releaseDate;
-    double duration;
+    int duration;
 }
